refactor: crée un dossier domaine/questions
