+++ site/source/domaine/questions/index.ts
@@ -1,0 +1,2 @@
+export { type QuestionPublicodes } from './QuestionPublicodes'
+export { type GroupeDeQuestionsPublicodes } from './GroupeDeQuestionsPublicodes'
